Fixed Tags & Dropdown menus
I fixed the tag links not working, but that created a bug where the nested sections would be displayed if you clicked on something like the "features" button which should simply drop down. If you wish to have nested sections listed when clicked, remove the "{{ if or (hasPrefix .URL "/blog/") (in .URL "tags") }}" and one of the "{{ ends }}" at the end.